Cache connections
If a user asks for a DB connection with the same options as those we've
seen before, send back the existing connection.

We were having issues where we kept opening up the same connection
again and again until we ran out of connections.

If a user closes a connection, we might have an issue here, but we'll
worry about that later.