Fix DbOpts::DbOptions#apply and tests
Sequel::Database#log_connection_execute should always be defined
correctly, or it is a bug in the database adapter.  It's a private
method, so respond_to? will return false unless you pass true as
the second argument.  Since it should always be defined, just
calling it with send seems best.

This broke the tests, which use an very brittle mocking design.
Switch the tests to use Sequel's mock database adapter, then nothing
needs to be mocked in the tests themselves, the tests are far more
robust, and also much easier to understand.