Add connection abstraction for Impala and PostgreSQL
This tries to abstract access parameters across the following
adapters (with recognized :adapter values in parentheses):

* jdbc/impala (:jdbc_impala)
* jdbc/hive2 (:jdbc_hive2)
* jdbc/postgresql (:jdbc_postgres)
* impala (:impala)
* postgres (:postgres)

It takes the following general connection options:

* :user
* :password
* :host
* :port
* :database

And for impala the following Kerberos related options:

* :realm
* :host_fqdn
* :principal

For the jdbc adapters it tries to construct the appropriate
JDBC URL, and for the impala adapter, it tries to construct the
correct connection hash.

This has had basic (non-Kerberos) testing so far on all adapters.
Specifically, the jdbc/hive2, jdbc/impala, and impala adapters
all work with the only change being the :adapter option.

This actually doesn't have any code specific to the :postgres
adapter, since the standard options work there.